补丁只改未修改的

+++ src/Install.php
@@ -28,7 +28,9 @@ class Install
         $app = new App();
         $patch_file = $app->getRootPath().'vendor/topthink/think-orm/src/db/PDOConnection.php';
         if(file_exists($patch_file)){
-            file_put_contents($patch_file, str_replace('protected $queryStartTime;', 'public $queryStartTime;', file_get_contents($patch_file)));
+            if(stripos(file_get_contents($patch_file), 'protected $queryStartTime;') !== false){
+                file_put_contents($patch_file, str_replace('protected $queryStartTime;', 'public $queryStartTime;', file_get_contents($patch_file)));
+            }
         }
     }
